requirements.txt
openai
pdf2image
opencv-python
pillow